requirements.txt
requests==2.28.1
python-dotenv==0.21.0
pathvalidate==2.5.2
bs4==0.0.1
jinja2==3.1.2